requirements.txt
lxml
beautifulsoup4
requests
python-frontmatter
markdown
termcolor
numpy
pillow
emoji
pymediainfo
